[enhancement] 1. 타입스크립트 적용 중2 (모듈 임포트 이슈 해결 중)
