+++ src/GlobalStyles.js
@@ -1,5 +1,5 @@
-import {createGlobalStyle} from 'styled-components';
-import palette from './lib/palette';
+import { createGlobalStyle } from 'styled-components';
+import palette from '../../dist/lib/palette';
 
 const GlobalStyles = createGlobalStyle`
 
@@ -27,5 +27,5 @@ h1,h2,h3,h4,h5,h6,ul {
 }
 
 li {list-style: none;}
-`
+`;
 export default GlobalStyles;
