--- src/lib/palette.js
@@ -1,10 +1,0 @@
-const palette = {
-    basicColor : '#eee6e3',
-    strongColor : '#f5b151',
-    wrapperColor :'#3c3f44',
-    backgroundColor: '#1F2124',
-    NewRankColor: '#F68657',
-    OldRankColor: '#c1c4bd'
-}
-
-export default palette;
