+++ src/reportWebVitals.tsx
@@ -1,0 +1,15 @@
+const reportWebVitals = (onPerfEntry?: Function | undefined) => {
+  if (onPerfEntry && onPerfEntry instanceof Function) {
+    require('web-vitals').then(
+      ({ getCLS, getFID, getFCP, getLCP, getTTFB }) => {
+        getCLS(onPerfEntry);
+        getFID(onPerfEntry);
+        getFCP(onPerfEntry);
+        getLCP(onPerfEntry);
+        getTTFB(onPerfEntry);
+      }
+    );
+  }
+};
+
+export default reportWebVitals;
