--- types/App.d.ts
@@ -1,2 +1,0 @@
-export default App;
-declare function App(): JSX.Element;
