--- types/common/Footer.d.ts
@@ -1,2 +1,0 @@
-export default Footer;
-declare function Footer(): JSX.Element;
