--- types/common/Header.d.ts
@@ -1,2 +1,0 @@
-export default Header;
-declare function Header(): JSX.Element;
