--- types/lib/deviceTheme.d.ts
@@ -1,12 +1,0 @@
-export default theme;
-declare namespace theme {
-    const mobileS: string;
-    const mobileSL: string;
-    const mobileM: string;
-    const mobileL: string;
-    const tabletS: string;
-    const tabletM: string;
-    const tabletL: string;
-    const laptop: string;
-    const desktop: string;
-}
