--- types/lib/palette.d.ts
@@ -1,9 +1,0 @@
-export default palette;
-declare namespace palette {
-    const basicColor: string;
-    const strongColor: string;
-    const wrapperColor: string;
-    const backgroundColor: string;
-    const NewRankColor: string;
-    const OldRankColor: string;
-}
