--- types/reportWebVitals.d.ts
@@ -1,2 +1,0 @@
-export default reportWebVitals;
-declare function reportWebVitals(onPerfEntry: any): void;
