+++ typings/App.d.ts
@@ -1,0 +1,2 @@
+declare const App: () => JSX.Element;
+export default App;
