+++ typings/GlobalStyles.d.ts
@@ -1,2 +1,2 @@
-export default GlobalStyles;
 declare const GlobalStyles: import("styled-components").GlobalStyleComponent<{}, import("styled-components").DefaultTheme>;
+export default GlobalStyles;
