+++ typings/common/Footer.d.ts
@@ -1,0 +1,2 @@
+declare const Footer: () => JSX.Element;
+export default Footer;
