+++ typings/common/Header.d.ts
@@ -1,0 +1,2 @@
+declare const Header: () => JSX.Element;
+export default Header;
