+++ typings/components/DropdownComponent.d.ts
@@ -1,4 +1,3 @@
-/// <reference types="react" />
 declare const DropdownComponent: ({ item1, item2, nation, nationHandler }: {
     item1: any;
     item2: any;
