+++ typings/components/Movie.d.ts
@@ -1,4 +1,3 @@
-/// <reference types="react" />
 import PropTypes from 'prop-types';
 declare const Movie: {
     ({ title, openDt, rank, rankOldAndNew, audiAcc }: {
