+++ typings/lib/deviceTheme/index.d.ts
@@ -1,0 +1,13 @@
+declare module 'theme';
+// {
+//     mobileS: string;
+//     mobileSL: string;
+//     mobileM: string;
+//     mobileL: string;
+//     tabletS: string;
+//     tabletM: string;
+//     tabletL: string;
+//     laptop: string;
+//     desktop: string;
+// };
+// export default theme;
