+++ typings/lib/palette/index.d.ts
@@ -1,0 +1,10 @@
+declare module 'palette';
+//  {
+//   basicColor: string;
+//   strongColor: string;
+//   wrapperColor: string;
+//   backgroundColor: string;
+//   NewRankColor: string;
+//   OldRankColor: string;
+// };
+// export default palette;
