+++ typings/reportWebVitals.d.ts
@@ -1,0 +1,2 @@
+declare const reportWebVitals: (onPerfEntry?: Function | undefined) => void;
+export default reportWebVitals;
